# [DNM] Try uncommenting `VectorProtocol` operator definitions.