Try uncommenting `VectorProtocol` operators to see if they break the stdlib:
- `VectorProtocol.+` (two variants)
- `VectorProtocol.+=`
- `VectorProtocol.-` (two variants)
- `VectorProtocol.-=`
- `VectorProtocol.*` (two variants)
- `VectorProtocol.*=`

If they do break the stdlib, we should consider publicizing/removing [internal
operator definitions from tensorflow/swift-apis](https://github.com/tensorflow/swift-apis/blob/f2acd6c4d1f14e7d0c1f9579dcc3631990e32602/Sources/TensorFlow/Operators/Math.swift#L223-L253), since it's surprising that the
operators aren't available to library users.