@dan-zheng 2020-02-13
Adding operator definitions to the stdlib cause a type-checking performance test to fail:
```
FAIL: Swift(macosx-x86_64) :: Sema/type_checker_perf/fast/rdar46541800.swift (7018 of 12900)
******************** TEST 'Swift(macosx-x86_64) :: Sema/type_checker_perf/fast/rdar46541800.swift' FAILED ********************
Script:
--
: 'RUN: at line 1';   /Users/swiftci/jenkins/workspace/swift-PR-TensorFlow-macOS/Ninja-ReleaseAssert+stdlib-Release/swift-macosx-x86_64/bin/swiftc -frontend -target x86_64-apple-macosx10.9  -module-cache-path '/Users/swiftci/jenkins/workspace/swift-PR-TensorFlow-macOS/Ninja-ReleaseAssert+stdlib-Release/swift-macosx-x86_64/swift-test-results/x86_64-apple-macosx10.9/clang-module-cache' -sdk '/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.15.sdk'  -swift-version 4  -ignore-module-source-info -typo-correction-limit 10  -typecheck -verify -disable-objc-attr-requires-foundation-module /Users/swiftci/jenkins/workspace/swift-PR-TensorFlow-macOS/swift/validation-test/Sema/type_checker_perf/fast/rdar46541800.swift -solver-expression-time-threshold=1
--
Exit Code: 1

Command Output (stderr):
--
/Users/swiftci/jenkins/workspace/swift-PR-TensorFlow-macOS/swift/validation-test/Sema/type_checker_perf/fast/rdar46541800.swift:19:13: error: unexpected error produced: the compiler is unable to type-check this expression in reasonable time; try breaking up the expression into distinct sub-expressions
            + (-descriptionTextView.firstBaselineOffsetFromTop - textInset.top + descriptionTextBaselineOffset).ceilingValue(scale: displayScale)
            ^
```

This suggests that type-checking performance is noticeably impacted.

Let's remove these [internals operator definitions from `tensorflow/swift-apis`](https://github.com/tensorflow/swift-apis/blob/f2acd6c4d1f14e7d0c1f9579dcc3631990e32602/Sources/TensorFlow/Operators/Math.swift#L223-L253) then, since it's surprising that the operators aren't available to library users.

cc @sgugger