@philipturner 2022-05-23
@BradLarson is it possible to check why this deteriorated type-checking performance in the first place? It seems like that would be a bug in the compiler. Also, why would the performance drop disappear when I add them back into S4TF as a standalone package?