@dan-zheng 2022-05-23
> @BradLarson is it possible to check why this deteriorated type-checking performance in the first place? It seems like that would be a bug in the compiler.

Some context:

The type-checking performance issue is due to Swift's [type checking approach](https://github.com/apple/swift/blob/main/docs/TypeChecker.md#approach), which needs to support a variety of language features: custom literal-convertible types and "function and operator overloading, subtyping, and constrained parametric polymorphism".

These links have some more context:
- https://github.com/apple/swift/blob/main/docs/TypeChecker.md
- https://www.cocoawithlove.com/blog/2016/07/12/type-checker-issues.html
- https://forums.swift.org/t/pitch-making-expression-type-checking-of-operator-expressions-fast/18037: a pitch to speed up operator type-checking that did not land (AFAIK) but explains context.

In short, operator type checking is expensive with these supported language features and with this approach, because every operator usage (e.g. `+`) introduces a disjunction constraint over all possible overloads of that operator (`Int.+`, `Float.+`, `Array.+`, etc), which quickly grows the number of [generated constraints](https://github.com/apple/swift/blob/main/docs/TypeChecker.md#constraint-generation).

> Also, why would the performance drop disappear when I add them back into S4TF as a standalone package?

Adding these `VectorProtocol.{+,+=,-,-=,*,*=}` operator definitions to a non-stdlib library should compile just fine. Type checking performance issues arise for the failing tests mentioned above (https://github.com/apple/swift/pull/29815#issuecomment-586027962), which stress-test operator type checking by creating complex expressions. Adding an extra overload of `+(_:_:)` causes generation of `n+1` disjunction constraints instead of `n` constraints, which super-linearly grows the constraint space and leads to `the compiler is unable to type-check this expression in reasonable time` diagnostics.