@dan-zheng 2022-05-23
As a library developer, it can be fine to add overloaded operators to your library – though if you notice type-checking performance issues, you could consider using new operators or using methods instead. And if you try to add operators to the stdlib (which was tracked by this issue), then you are more likely to encounter type checking performance issues because the stdlib is implicitly imported by every program.