@philipturner 2022-05-23
Workaround removed in https://github.com/s4tf/s4tf/pull/13. It looks like there is no deterioration in compile-time performance, at least when compiling the `TensorFlow` Swift module. Here are the build times before and after adding the operators back in. I cleared the Swift package build folder before each run.

Before adding back in:
```
Build complete! (35.13s) - TensorFlow
Build complete! (21.51s) - TensorFlowTests
```
After adding back in:
```
Build complete! (32.91s) - TensorFlow
Build complete! (21.57s) - TensorFlowTests
```
After removing again to test the questionable decrease in compile time from 35.13 to 32.91 seconds:
```
Build complete! (32.60s) - TensorFlow
Build complete! (21.72s) - TensorFlowTests
```

<details>
<summary>Full build log</summary>

```
(base) philipturner@M1-Max-MacBook-Pro building-tensorflow % bash build2.sh --skip-ctensorflow
objc[61212]: Class _TtCs25CheckedContinuationCanary is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de670) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1065288c8). One of the two will be used. Which one is undefined.
objc[61212]: Class _TtCs17DispatchQueueShim is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de700) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x106528958). One of the two will be used. Which one is undefined.
objc[61212]: Class _TtCs9MainActor is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de790) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1065289f0). One of the two will be used. Which one is undefined.
warning: Usage of /Users/philipturner/Library/org.swift.swiftpm/collections.json has been deprecated. Please delete it and use the new /Users/philipturner/Library/org.swift.swiftpm/configuration/collections.json instead.
Fetching https://github.com/apple/swift-numerics from cache
Fetched https://github.com/apple/swift-numerics (0.69s)
Fetching https://github.com/pvieito/PythonKit.git from cache
Fetched https://github.com/pvieito/PythonKit.git (0.77s)
Creating working copy for https://github.com/apple/swift-numerics
Working copy of https://github.com/apple/swift-numerics resolved at main
Creating working copy for https://github.com/pvieito/PythonKit.git
Working copy of https://github.com/pvieito/PythonKit.git resolved at master
Building for debugging...
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
[129/129] Linking libx10_optimizers_optimizer.dylib
Build complete! (35.13s)
objc[61995]: Class _TtCs25CheckedContinuationCanary is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de670) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x103fd08c8). One of the two will be used. Which one is undefined.
objc[61995]: Class _TtCs17DispatchQueueShim is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de700) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x103fd0958). One of the two will be used. Which one is undefined.
objc[61995]: Class _TtCs9MainActor is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de790) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x103fd09f0). One of the two will be used. Which one is undefined.
warning: Usage of /Users/philipturner/Library/org.swift.swiftpm/collections.json has been deprecated. Please delete it and use the new /Users/philipturner/Library/org.swift.swiftpm/configuration/collections.json instead.
Building for debugging...
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
[43/43] Linking TensorFlowPackageTests
Build complete! (21.51s)
2022-05-23 19:22:11.426 xctest[62100:1434083] The bundle “TensorFlowPackageTests.xctest” couldn’t be loaded. Try reinstalling the bundle.
2022-05-23 19:22:11.426 xctest[62100:1434083] (dlopen(/Users/philipturner/Documents/building-tensorflow/s4tf/.build/arm64-apple-macosx/debug/TensorFlowPackageTests.xctest/Contents/MacOS/TensorFlowPackageTests, 0x0109): Symbol not found: _$sSa16_DifferentiationAA14DifferentiableRzlE0B4ViewV4baseSayxGvg
  Referenced from: /Users/philipturner/Documents/building-tensorflow/s4tf/.build/arm64-apple-macosx/debug/TensorFlowPackageTests.xctest/Contents/MacOS/TensorFlowPackageTests
  Expected in: /usr/lib/swift/libswift_Differentiation.dylib)
Test Suite 'All tests' started at 2022-05-23 19:22:11.925
Test Suite 'TensorFlowPackageTests.xctest' started at 2022-05-23 19:22:11.926
Test Suite 'BasicOperatorTests' started at 2022-05-23 19:22:11.926
...
Test Suite 'All tests' failed at 2022-05-23 19:22:18.910.
	 Executed 477 tests, with 4 failures (0 unexpected) in 6.977 (6.985) seconds
(base) philipturner@M1-Max-MacBook-Pro building-tensorflow % bash build2.sh --skip-ctensorflow
objc[62113]: Class _TtCs25CheckedContinuationCanary is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de670) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1068d08c8). One of the two will be used. Which one is undefined.
objc[62113]: Class _TtCs17DispatchQueueShim is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de700) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1068d0958). One of the two will be used. Which one is undefined.
objc[62113]: Class _TtCs9MainActor is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de790) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1068d09f0). One of the two will be used. Which one is undefined.
warning: Usage of /Users/philipturner/Library/org.swift.swiftpm/collections.json has been deprecated. Please delete it and use the new /Users/philipturner/Library/org.swift.swiftpm/configuration/collections.json instead.
Fetching https://github.com/apple/swift-numerics from cache
Fetched https://github.com/apple/swift-numerics (0.66s)
Fetching https://github.com/pvieito/PythonKit.git from cache
Fetched https://github.com/pvieito/PythonKit.git (0.70s)
Creating working copy for https://github.com/apple/swift-numerics
Working copy of https://github.com/apple/swift-numerics resolved at main
Creating working copy for https://github.com/pvieito/PythonKit.git
Working copy of https://github.com/pvieito/PythonKit.git resolved at master
Building for debugging...
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
[129/129] Linking libx10_optimizers_optimizer.dylib
Build complete! (32.91s)
objc[62849]: Class _TtCs25CheckedContinuationCanary is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de670) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1028c08c8). One of the two will be used. Which one is undefined.
objc[62849]: Class _TtCs17DispatchQueueShim is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de700) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1028c0958). One of the two will be used. Which one is undefined.
objc[62849]: Class _TtCs9MainActor is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de790) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x1028c09f0). One of the two will be used. Which one is undefined.
warning: Usage of /Users/philipturner/Library/org.swift.swiftpm/collections.json has been deprecated. Please delete it and use the new /Users/philipturner/Library/org.swift.swiftpm/configuration/collections.json instead.
Building for debugging...
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
[43/43] Linking TensorFlowPackageTests
Build complete! (21.57s)
2022-05-23 19:24:07.180 xctest[62954:1436844] The bundle “TensorFlowPackageTests.xctest” couldn’t be loaded. Try reinstalling the bundle.
2022-05-23 19:24:07.180 xctest[62954:1436844] (dlopen(/Users/philipturner/Documents/building-tensorflow/s4tf/.build/arm64-apple-macosx/debug/TensorFlowPackageTests.xctest/Contents/MacOS/TensorFlowPackageTests, 0x0109): Symbol not found: _$sSa16_DifferentiationAA14DifferentiableRzlE0B4ViewV4baseSayxGvg
  Referenced from: /Users/philipturner/Documents/building-tensorflow/s4tf/.build/arm64-apple-macosx/debug/TensorFlowPackageTests.xctest/Contents/MacOS/TensorFlowPackageTests
  Expected in: /usr/lib/swift/libswift_Differentiation.dylib)
Test Suite 'All tests' started at 2022-05-23 19:24:07.831
Test Suite 'TensorFlowPackageTests.xctest' started at 2022-05-23 19:24:07.831
Test Suite 'BasicOperatorTests' started at 2022-05-23 19:24:07.831
...
Test Suite 'All tests' failed at 2022-05-23 19:24:14.782.
	 Executed 477 tests, with 4 failures (0 unexpected) in 6.944 (6.951) seconds
(base) philipturner@M1-Max-MacBook-Pro building-tensorflow % cd s4tf
(base) philipturner@M1-Max-MacBook-Pro s4tf % git add .
(base) philipturner@M1-Max-MacBook-Pro s4tf % git commit -m "Add VectorProtocol operators"
[remove-pr29815-workaround 8b295b6] Add VectorProtocol operators
 2 files changed, 34 insertions(+), 36 deletions(-)
(base) philipturner@M1-Max-MacBook-Pro s4tf % git branch
  bump-to-tf2.9
  main
  regenerate-raw-ops
* remove-pr29815-workaround
  reorganize-tensor-module
  update-manual-xla-docs
(base) philipturner@M1-Max-MacBook-Pro s4tf % git push philipturner remove-pr29815-workaround
Enumerating objects: 13, done.
Counting objects: 100% (13/13), done.
Delta compression using up to 10 threads
Compressing objects: 100% (7/7), done.
Writing objects: 100% (7/7), 771 bytes | 771.00 KiB/s, done.
Total 7 (delta 6), reused 0 (delta 0), pack-reused 0
remote: Resolving deltas: 100% (6/6), completed with 6 local objects.
To https://github.com/philipturner/s4tf.git
   b212161..8b295b6  remove-pr29815-workaround -> remove-pr29815-workaround
(base) philipturner@M1-Max-MacBook-Pro s4tf % git checkout maion
error: pathspec 'maion' did not match any file(s) known to git
(base) philipturner@M1-Max-MacBook-Pro s4tf % git checkout main
Switched to branch 'main'
Your branch is up to date with 'origin/main'.
(base) philipturner@M1-Max-MacBook-Pro s4tf % bash build2.sh --skip-ctensorflow              
bash: build2.sh: No such file or directory
(base) philipturner@M1-Max-MacBook-Pro s4tf % cd ../
(base) philipturner@M1-Max-MacBook-Pro building-tensorflow % bash build2.sh --skip-ctensorflow
objc[63044]: Class _TtCs25CheckedContinuationCanary is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de670) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x10448c8c8). One of the two will be used. Which one is undefined.
objc[63044]: Class _TtCs17DispatchQueueShim is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de700) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x10448c958). One of the two will be used. Which one is undefined.
objc[63044]: Class _TtCs9MainActor is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de790) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x10448c9f0). One of the two will be used. Which one is undefined.
warning: Usage of /Users/philipturner/Library/org.swift.swiftpm/collections.json has been deprecated. Please delete it and use the new /Users/philipturner/Library/org.swift.swiftpm/configuration/collections.json instead.
Fetching https://github.com/apple/swift-numerics from cache
Fetched https://github.com/apple/swift-numerics (0.67s)
Fetching https://github.com/pvieito/PythonKit.git from cache
Fetched https://github.com/pvieito/PythonKit.git (0.64s)
Creating working copy for https://github.com/pvieito/PythonKit.git
Working copy of https://github.com/pvieito/PythonKit.git resolved at master
Creating working copy for https://github.com/apple/swift-numerics
Working copy of https://github.com/apple/swift-numerics resolved at main
Building for debugging...
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
[129/129] Linking libx10_optimizers_optimizer.dylib
Build complete! (32.60s)
objc[63783]: Class _TtCs25CheckedContinuationCanary is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de670) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x106e288c8). One of the two will be used. Which one is undefined.
objc[63783]: Class _TtCs17DispatchQueueShim is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de700) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x106e28958). One of the two will be used. Which one is undefined.
objc[63783]: Class _TtCs9MainActor is implemented in both /usr/lib/swift/libswift_Concurrency.dylib (0x2282de790) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-05-04-a.xctoolchain/usr/lib/swift/macosx/libswift_Concurrency.dylib (0x106e289f0). One of the two will be used. Which one is undefined.
warning: Usage of /Users/philipturner/Library/org.swift.swiftpm/collections.json has been deprecated. Please delete it and use the new /Users/philipturner/Library/org.swift.swiftpm/configuration/collections.json instead.
Building for debugging...
ld: warning: dylib (/Users/philipturner/Documents/building-tensorflow/Library/tensorflow-2.9.0/usr/lib/libx10.dylib) was built for newer macOS version (12.3) than being linked (11.0)
[43/43] Linking TensorFlowPackageTests
Build complete! (21.72s)
2022-05-23 19:32:33.689 xctest[63888:1442633] The bundle “TensorFlowPackageTests.xctest” couldn’t be loaded. Try reinstalling the bundle.
2022-05-23 19:32:33.690 xctest[63888:1442633] (dlopen(/Users/philipturner/Documents/building-tensorflow/s4tf/.build/arm64-apple-macosx/debug/TensorFlowPackageTests.xctest/Contents/MacOS/TensorFlowPackageTests, 0x0109): Symbol not found: _$sSa16_DifferentiationAA14DifferentiableRzlE0B4ViewV4baseSayxGvg
  Referenced from: /Users/philipturner/Documents/building-tensorflow/s4tf/.build/arm64-apple-macosx/debug/TensorFlowPackageTests.xctest/Contents/MacOS/TensorFlowPackageTests
  Expected in: /usr/lib/swift/libswift_Differentiation.dylib)
Test Suite 'All tests' started at 2022-05-23 19:32:34.210
Test Suite 'TensorFlowPackageTests.xctest' started at 2022-05-23 19:32:34.210
Test Suite 'BasicOperatorTests' started at 2022-05-23 19:32:34.210
...
^CCaught interrupt signal, exiting...
```

</details>


`build2.sh` is based on https://gist.github.com/philipturner/7aa063af04277d463c14168275878511